LPS-51081 Move stuff out
